Feature/add friend (#111)
This is a very basic implementation of an add friend feature. The user inputs an email and the backend will search to see if that user exists. If the user does exist, that user is added as a MongoDB ref into the user's friends list.

The user's friend's list is populated with friends' email, name and avatar whenever the user adds a friend, logs in, or has their token refreshed.

On a future implementation it could be good to allow the user to add multiple friends at a time and send friends an email invitation to join the app if they are not already registered.

**Link to Issue**:

-   Fixes/Closes #79 

**Readiness**:

-   [x] Avoid duplicated logic
-   [ ] Unit test component functionalities
-   [x] Use semantic HTML tags
-   [x] Use [semantic commits](https://www.conventionalcommits.org/en/v1.0.0-beta.2/)
-   [x] Connect this Pull Request to an issue