Feature/113: Public profile route (#114)
Implemented route that will expose user profile data to the public, and updated user tests. 

An idea for a feature or fixing security issue: Allow users to decide what data they want to display as public and what should be hidden. Currently, most of the user data (which is not much) is exposed.

**Link to Issue**:

-   Fixes/Closes #113 

**Readiness**:

-   [x] Avoid duplicated logic
-   [x] Unit test component functionalities
-   [x] Use semantic HTML tags
-   [x] Use [semantic commits](https://www.conventionalcommits.org/en/v1.0.0-beta.2/)
-   [x] Connect this Pull Request to an issue