Test/119: Login and register unit tests (#134)
Unit tests for the Login and Register components, as well as the login and register Redux action creators.

There is a `createTestStore()` function in test utils which can be used to create an actual redux store for testing. This is useful for testing state changes on the store.

This PR also adds the `redux-mock-store` library. This is useful to test whether particular actions are called on the store, but because it is a mock it does not hold actual state.

**Link to Issue**:

-   Fixes/Closes #119 

**Readiness**:

-   [x] Avoid duplicated logic
-   [x] Unit test component functionalities
-   [x] Use semantic HTML tags
-   [x] Use [semantic commits](https://www.conventionalcommits.org/en/v1.0.0-beta.2/)
-   [x] Connect this Pull Request to an issue